feat: Implement delete operation, remove redundant code, fix in16 function signature

+++ src/lib-header/portio.h
@@ -21,7 +21,7 @@ void out16(uint16_t port, uint16_t data);
  */
 uint8_t in(uint16_t port);
 
-uint8_t in16(uint16_t port);
+uint16_t in16(uint16_t port);
 
 
 #endif
